final final

+++ Assets/Scripts/Obstacle.cs
@@ -16,12 +16,12 @@ public class Obstacle : MonoBehaviour
         Transform needle2 = gameObject.transform.GetChild(3);
         Transform needle3 = gameObject.transform.GetChild(4);
 
-        float x1 = needle1.position.x + Random.Range(XMin, XMax);
-        float y1 = needle1.position.y + Random.Range(YMin, YMax);
-        float x2 = needle2.position.x + Random.Range(XMin, XMax);
-        float y2 = needle2.position.y + Random.Range(YMin, YMax);
-        float x3 = needle3.position.x + Random.Range(XMin, XMax);
-        float y3 = needle3.position.y + Random.Range(YMin, YMax);
+        float x1 = needle1.position.x + (int)Random.Range(XMin, XMax);
+        float y1 = needle1.position.y + (int)Random.Range(YMin, YMax);
+        float x2 = needle2.position.x + (int)Random.Range(XMin, XMax);
+        float y2 = needle2.position.y + (int)Random.Range(YMin, YMax);
+        float x3 = needle3.position.x + (int)Random.Range(XMin, XMax);
+        float y3 = needle3.position.y + (int)Random.Range(YMin, YMax);
 
         needle1.position = new Vector3(x1, y1, 1);
         needle2.position = new Vector3(x2, y2, 1);
